+++ SpringMicroservices/universitySE/src/main/java/com/example/universitySE/domain/Exam.java
@@ -1,5 +1,5 @@
 package com.example.universitySE.domain;
-// Generated 8-feb-2019 18.41.41 by Hibernate Tools 5.2.0.Final
+// Generated 14-feb-2019 21.58.38 by Hibernate Tools 5.2.0.Final
 
 
 import javax.persistence.Column;
@@ -17,12 +17,12 @@ import javax.persistence.Table;
 public class Exam  implements java.io.Serializable {
 
 
-     private long id;
+     private int id;
 
     public Exam() {
     }
 
-    public Exam(long id) {
+    public Exam(int id) {
        this.id = id;
     }
    
@@ -30,11 +30,11 @@ public class Exam  implements java.io.Serializable {
 
     
     @Column(name="id", unique=true, nullable=false)
-    public long getId() {
+    public int getId() {
         return this.id;
     }
     
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
