+++ SpringMicroservices/universitySE/src/main/java/com/example/universitySE/domain/PersonType.java
@@ -1,5 +1,5 @@
 package com.example.universitySE.domain;
-// Generated 8-feb-2019 18.41.41 by Hibernate Tools 5.2.0.Final
+// Generated 14-feb-2019 21.58.38 by Hibernate Tools 5.2.0.Final
 
 
 import javax.persistence.*;
@@ -16,7 +16,7 @@ import static javax.persistence.GenerationType.IDENTITY;
 public class PersonType  implements java.io.Serializable {
 
 
-     private long id;
+     private Integer id;
      private String name;
 
     public PersonType() {
@@ -30,11 +30,11 @@ public class PersonType  implements java.io.Serializable {
 
     
     @Column(name="id", unique=true, nullable=false)
-    public long getId() {
+    public Integer getId() {
         return this.id;
     }
     
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
