+++ SpringMicroservices/universitySE/src/main/java/com/example/universitySE/domain/Professor.java
@@ -1,5 +1,5 @@
 package com.example.universitySE.domain;
-// Generated 8-feb-2019 18.41.41 by Hibernate Tools 5.2.0.Final
+// Generated 14-feb-2019 21.58.38 by Hibernate Tools 5.2.0.Final
 
 
 import javax.persistence.*;
@@ -15,7 +15,7 @@ import java.util.Date;
 public class Professor  implements java.io.Serializable {
 
 
-     private long id;
+     private int id;
      private String firstName;
      private String lastName;
      private String biography;
@@ -27,7 +27,7 @@ public class Professor  implements java.io.Serializable {
     }
 
 	
-    public Professor(long id, String firstName, String lastName, String receptionTime, int subject, Date dateOfBirth) {
+    public Professor(int id, String firstName, String lastName, String receptionTime, int subject, Date dateOfBirth) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -35,7 +35,7 @@ public class Professor  implements java.io.Serializable {
         this.subject = subject;
         this.dateOfBirth = dateOfBirth;
     }
-    public Professor(long id, String firstName, String lastName, String biography, String receptionTime, int subject, Date dateOfBirth) {
+    public Professor(int id, String firstName, String lastName, String biography, String receptionTime, int subject, Date dateOfBirth) {
        this.id = id;
        this.firstName = firstName;
        this.lastName = lastName;
@@ -49,11 +49,11 @@ public class Professor  implements java.io.Serializable {
 
     
     @Column(name="id", unique=true, nullable=false)
-    public long getId() {
+    public int getId() {
         return this.id;
     }
     
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
