+++ SpringMicroservices/universitySE/src/main/java/com/example/universitySE/domain/Secretary.java
@@ -1,5 +1,5 @@
 package com.example.universitySE.domain;
-// Generated 8-feb-2019 18.41.41 by Hibernate Tools 5.2.0.Final
+// Generated 14-feb-2019 21.58.38 by Hibernate Tools 5.2.0.Final
 
 
 import javax.persistence.Column;
@@ -17,14 +17,14 @@ import javax.persistence.Table;
 public class Secretary  implements java.io.Serializable {
 
 
-     private long id;
+     private int id;
      private int faculty;
      private String venue;
 
     public Secretary() {
     }
 
-    public Secretary(long id, int faculty, String venue) {
+    public Secretary(int id, int faculty, String venue) {
        this.id = id;
        this.faculty = faculty;
        this.venue = venue;
@@ -34,11 +34,11 @@ public class Secretary  implements java.io.Serializable {
 
     
     @Column(name="id", unique=true, nullable=false)
-    public long getId() {
+    public int getId() {
         return this.id;
     }
     
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
