+++ SpringMicroservices/universitySE/src/main/java/com/example/universitySE/domain/StudyCourse.java
@@ -1,5 +1,5 @@
 package com.example.universitySE.domain;
-// Generated 8-feb-2019 18.41.41 by Hibernate Tools 5.2.0.Final
+// Generated 14-feb-2019 21.58.38 by Hibernate Tools 5.2.0.Final
 
 
 import javax.persistence.*;
@@ -16,14 +16,14 @@ import static javax.persistence.GenerationType.IDENTITY;
 public class StudyCourse  implements java.io.Serializable {
 
 
-     private long id;
+     private Integer id;
      private String name;
-     private long faculty;
+     private int faculty;
 
     public StudyCourse() {
     }
 
-    public StudyCourse(String name, long faculty) {
+    public StudyCourse(String name, int faculty) {
        this.name = name;
        this.faculty = faculty;
     }
@@ -32,11 +32,11 @@ public class StudyCourse  implements java.io.Serializable {
 
     
     @Column(name="id", unique=true, nullable=false)
-    public long getId() {
+    public Integer getId() {
         return this.id;
     }
     
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -52,11 +52,11 @@ public class StudyCourse  implements java.io.Serializable {
 
     
     @Column(name="faculty", nullable=false)
-    public long getFaculty() {
+    public int getFaculty() {
         return this.faculty;
     }
     
-    public void setFaculty(long faculty) {
+    public void setFaculty(int faculty) {
         this.faculty = faculty;
     }
