+++ SpringMicroservices/universitySE/src/main/java/com/example/universitySE/dtos/PersonDTO.java
@@ -38,7 +38,7 @@ public class PersonDTO {
     private Date dateOfBirth;
 
     @Getter @Setter
-    private long badgeNumber;
+    private int badgeNumber;
 
     @Getter @Setter
     private int studyCourse;
